Update ng-context-menu.js
The position fixed allows to place the div content menu anywhere.
I take the menuElement only on click, so the dom/html is ready, I use your directive in another directive and with your code it doesn't find the menuElement because your directive is executing before my directive is completed. Do a test with your directive into another.